extract social action handlers into new useSocialActions hook and update layout styles
