+++ containers/frontend/web/src/features/social/social-variants/pending-buttons.tsx
@@ -1,10 +1,10 @@
 import { useTranslations } from 'next-intl';
 import { IconButton } from '@/components/composites/icon-button/icon-button';
-import { useSocialData } from '../hooks/use-social-data';
+import { useSocialActions } from '../hooks/use-social-actions';
 
 export function PendingButton({ friendshipId }: { friendshipId: string }) {
   const tActions = useTranslations('features.social.actions');
-  const { handleDelete } = useSocialData();
+  const { handleDelete } = useSocialActions();
 
   return (
     <IconButton
