+++ containers/frontend/web/src/features/social/social-variants/request-buttons.tsx
@@ -1,10 +1,10 @@
 import { useTranslations } from 'next-intl';
 import { IconButton } from '@/components/composites/icon-button/icon-button';
-import { useSocialData } from '../hooks/use-social-data';
+import { useSocialActions } from '../hooks/use-social-actions';
 
 export function RequestButtons({ friendshipId }: { friendshipId: string }) {
   const tActions = useTranslations('features.social.actions');
-  const { handleResponse } = useSocialData();
+  const { handleResponse } = useSocialActions();
 
   return (
     <>
